+++ lib/another_flushbar.dart
@@ -1,0 +1,5 @@
+library another_flushbar;
+
+export 'flushbar.dart';
+export 'flushbar_helper.dart';
+export 'flushbar_remote.dart';